Separating out construction into a codegen crate.
