+++ .gitignore
@@ -1,2 +1,4 @@
 /target
 Cargo.lock
+*/**/target
+*/**/Cargo.lock
